runs but into errors
