+++ NameIt.py
@@ -12,16 +12,13 @@ import unicodedata
 import magic
 
 
-
 # 4 Dynamic module loading
 from typing import Optional, Any
 
-import rich.traceback
-from loguru import logger
-from rich import print as rprint
-from rich.console import Console
 
-rich.traceback.install(show_locals=True)
+from utils.unified_logger import console, logger
+from NameItCrossRef import extract_doi_from_pdf
+
 
 
 class ModuleLoader:
@@ -113,16 +110,6 @@ def load_layoutlm_modules() -> bool:
         torch = loader.loaded_modules['torch']
 
     return success
-
-
-# Set up logging with loguru
-logger.remove()  # Remove the default logger
-logger.add(sys.stderr, level="INFO")
-
-# Set up rich console
-console = Console()
-
-
 
 
 def is_pdf_file(file_path: str) -> bool:
@@ -296,42 +283,6 @@ def extract_info_from_pdf(pdf_path):
         info[key] = info[key].strip()
 
     return info
-
-
-# Opening PDF file, reading the first page and extracting DOI with a very common pattern
-def extract_metadata_from_pdf(pdf_file):
-    try:
-        pdf_document = fitz.open(pdf_file)
-        first_page = pdf_document[0]
-        text = first_page.get_text("text")
-        doi_pattern = r'10[.][\d.]{1,15}\/[-._;:()\/A-Za-z0-9<>]+[A-Za-z0-9]'
-        doi_match = re.search(doi_pattern, text)
-        if doi_match:
-            doi = doi_match.group()
-            logger.info("Extracting DOI from file")
-            return fetch_metadata_by_doi(doi)
-        else:
-            return None
-    except re.error as e:
-        logger.error(f"Error with DOI regex pattern: {e}")
-    except Exception as e:
-        logger.error(f"Unexpected error extracting metadata from PDF: {e}")
-    return None
-
-
-# Using Crossref API to match the extracted DOI
-def fetch_metadata_by_doi(doi):
-    try:
-        cr = Crossref()
-        metadata = cr.works(doi)
-        logger.info("Extracting metadata through crossref.org")
-        rprint(metadata)  # Use rich to print the metadata
-        return metadata
-    except Crossref.exceptions.CrossrefError as e:
-        logger.error(f"Crossref API error: {e}")
-    except Exception as e:
-        logger.error(f"Unexpected error fetching metadata by DOI: {e}")
-    return None
 
 
 # Validate that the author field is a string
@@ -610,7 +561,7 @@ def process_folder_or_file(path: str, args: argparse.Namespace) -> None:
         # Verify at least one processing method is specified
         if args.use_pdf_metadata or args.use_crossref or args.use_layoutlmv3:
             pdf_file = path
-            metadata = extract_metadata_from_pdf(pdf_file)
+            metadata = extract_doi_from_pdf(pdf_file)
         else:
             console.print("[red]Method not known.[/red]")
             console.print("[blue]Are you sure you don't want to use pdf metadata? "
